backend/requirements.txt
aiosmtplib==4.0.2
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cramjam==2.11.0
cryptography==46.0.3
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fonttools==4.57.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.14.0
httptools==0.7.1
idna==3.11
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mongo==0.2.0
motor==3.7.1
numpy==2.0.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.4
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.0
python-jose==3.5.0
python-multipart==0.0.20
python-snappy==0.7.3
python-socketio==5.13.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
trimesh==4.6.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.2.0
zipp==3.21.0
